pyproject.toml
[project]
name = "afriagent"
version = "0.1.0"
description = "Production-grade AI customer support agent for African businesses"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "AfriAgent Team" }]
readme = "README.md"
dependencies = [
    # Core
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0",
    "structlog>=23.0",
    "python-dotenv>=1.0",
    
    # Async
    "anyio>=4.0",
    
    # HTTP
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    
    # Redis
    "redis[hiredis]>=5.0",
    
    # Postgres
    "asyncpg>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    
    # Vector DB
    "qdrant-client>=1.8",
    
    # LLM
    "openai>=1.12",
    "anthropic>=0.25",
    
    # Observability
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "prometheus-client>=0.20",
    
    # Validation
    "phonenumbers>=8.13",
]

[project.optional-dependencies]
telegram = ["python-telegram-bot>=21.0"]
ollama = ["ollama>=0.1"]
dev = [
    "ruff>=0.3",
    "mypy>=1.8",
    "pre-commit>=3.6",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "fakeredis>=2.21",
]

[project.scripts]
afriagent = "afriagent.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/afriagent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: unit tests",
    "integration: tests needing redis/postgres/qdrant",
    "e2e: end-to-end tests",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
